Fix narrowing between types

+++ engine/animations/joints/Joint.cpp
@@ -5,7 +5,7 @@
 #include <iostream>
 #include "Joint.h"
 
-Joint::Joint(std::string name, int index, const glm::mat4 &bindLocalTransform)
+Joint::Joint(std::string name, size_t index, const glm::mat4 &bindLocalTransform)
 		: m_Name(std::move(name)), m_Index(index), m_LocalBindTransform(glm::transpose(bindLocalTransform))
 {
 }
